add funcionality login and person online

+++ src/app/services/peticiones.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
-
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { environment } from '../../environments/environment';
+import { Router } from '@angular/router';
 
 @Injectable({
   providedIn: 'root'
@@ -11,24 +11,41 @@ export class PeticionesService {
 
   url = environment.MICROSERVICE_URL;
 
-  constructor(private http: HttpClient) {
-
+  constructor(private http: HttpClient, private router: Router) {
   }
 
   get(uri: string, headers?: HttpHeaders): Observable<any> {
+    if (this.authorizationToken()) {
     return this.http.get(this.url + uri, { observe: 'response', headers });
+    }
   }
 
   post(uri: string, data: any, headers?: HttpHeaders): Observable<any> {
-    console.log(this.url + uri);
+    if (this.authorizationToken()) {
     return this.http.post(this.url + uri, data, { observe: 'response', headers });
+    }
   }
 
   put(uri: string, data: any, headers?: HttpHeaders): Observable<any> {
+    if (this.authorizationToken()) {
     return this.http.put(this.url + uri, data, { observe: 'response', headers });
+    }
   }
 
   delete(uri: string, headers?: HttpHeaders): Observable<any> {
+    if (this.authorizationToken()) {
     return this.http.delete(this.url + uri, { observe: 'response', headers });
+    }
   }
+
+  private authorizationToken(): boolean {
+    const token = localStorage.getItem('jwt');
+    const usuario = localStorage.getItem('user');
+
+    if (token === '' || usuario === '') {
+      this.router.navigate(['login']);
+    }
+    return true;
+  }
+
 }
